requirements.txt
elasticsearch>=7.17.0,<8.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
openai>=1.0.0
